Extract reusable Feishu connector module
