+++ lib/connector-driver-transports.mjs
@@ -1,8 +1,11 @@
-import { createHash } from 'crypto';
 import { sendOutboundEmail } from './agent-mail-outbound.mjs';
 import { renderOutboundChatMessageText, normalizeConnectorSendResult } from './connector-driver.mjs';
 import { sendWeChatText } from '../scripts/wechat-connector.mjs';
-import { buildFeishuPostContent } from '../scripts/feishu-connector.mjs';
+import {
+  buildFeishuApiUuid,
+  buildFeishuPostContent,
+  isFeishuTopicSummary,
+} from '../connectors/feishu/index.mjs';
 
 function trimString(value) {
   return typeof value === 'string' ? value.trim() : '';
@@ -14,36 +17,6 @@ function firstNonEmpty(...values) {
     if (trimmed) return trimmed;
   }
   return '';
-}
-
-function normalizeFeishuMode(value) {
-  return trimString(value).toLowerCase().replace(/[_\s-]+/g, '_');
-}
-
-function isFeishuTopicSummary(summary) {
-  const chatType = normalizeFeishuMode(summary?.chatType);
-  const groupMessageType = normalizeFeishuMode(summary?.groupMessageType);
-  const chatMode = normalizeFeishuMode(summary?.chatMode);
-  return chatType === 'topic'
-    || groupMessageType === 'thread'
-    || groupMessageType === 'topic'
-    || chatMode === 'thread'
-    || chatMode === 'topic'
-    || !!trimString(summary?.threadId)
-    || !!trimString(summary?.rootId);
-}
-
-function buildFeishuApiUuid(value) {
-  const seed = trimString(value) || `${Date.now()}`;
-  const normalized = seed
-    .replace(/[^a-zA-Z0-9_-]+/g, '_')
-    .replace(/_+/g, '_')
-    .replace(/^_+|_+$/g, '');
-  if (normalized && normalized.length <= 64) {
-    return normalized;
-  }
-  const digest = createHash('sha1').update(seed).digest('hex').slice(0, 32);
-  return `rl_${digest}`;
 }
 
 function summarizeEmailExternalId(result) {
