+++ package.json
@@ -85,7 +85,7 @@
     "test:restart-gate": "node scripts/run-with-clean-instance-env.mjs npm run test:restart-gate:raw",
     "test:restart-gate:raw": "node tests/test-chat-static-assets.mjs && node tests/test-chat-split-frontend-smoke.mjs && node tests/test-build-info-versioning.mjs && node tests/test-static-asset-routing.mjs && node tests/test-chat-loader-versioned-assets.mjs && node tests/test-chat-build-update-indicator.mjs && node tests/test-source-runtime-only.mjs && node tests/test-cli-release-removed.mjs && node tests/test-cli-legacy-command-removed.mjs",
     "test:smoke": "node scripts/run-with-clean-instance-env.mjs npm run test:smoke:raw",
-    "test:smoke:raw": "node tests/test-chat-session-state-model.mjs && node tests/test-chat-session-store.mjs && node tests/test-chat-composer-store.mjs && node tests/test-chat-static-assets.mjs && node tests/test-session-naming.mjs && node tests/test-session-connector-naming.mjs && node tests/test-cloudflared-config.mjs && node tests/test-tunnel-diagnostics.mjs && node tests/test-history-index-contract.mjs && node tests/test-session-route-utils.mjs && node tests/test-apps-builtins.mjs && node tests/test-session-app-scope.mjs && node tests/test-session-external-trigger-refresh.mjs && node tests/test-session-follow-up-queue.mjs && node tests/test-session-label-prompt-context.mjs && node tests/test-auto-compaction.mjs && node tests/test-agent-mailbox.mjs && node tests/test-agent-mail-worker.mjs && node tests/test-embedded-mail-worker.mjs && node tests/test-agent-mail-reply.mjs && node tests/test-agent-mail-command.mjs && node tests/test-agent-mail-cloudflare-routing.mjs && node tests/test-mailbox-runtime-registry.mjs && node tests/test-mailbox-ingress-self-check.mjs && node tests/test-feishu-connector.mjs && node tests/test-wechat-connector.mjs && node tests/test-whatsapp-business-connector.mjs && node tests/test-connector-surface-public-paths.mjs && node tests/test-connector-surface-health.mjs && node tests/test-guest-wechat-connector-startup.mjs && node tests/test-feishu-ops.mjs && node tests/test-voice-connector.mjs && node tests/test-remote-capability-monitor.mjs && node tests/test-github-ci-auto-repair.mjs && node tests/test-github-ci-auto-repair-runner.mjs && node tests/test-generate-token-preserves-password.mjs && node tests/test-chat-instance-sync.mjs && node tests/test-guest-instance-command.mjs && node tests/test-normalize-owner-instance.mjs && node tests/test-source-runtime-only.mjs && node tests/test-cli-release-removed.mjs && node tests/test-oversized-files-report.mjs && npm run test:merge-safety",
+    "test:smoke:raw": "node tests/test-chat-session-state-model.mjs && node tests/test-chat-session-store.mjs && node tests/test-chat-composer-store.mjs && node tests/test-chat-static-assets.mjs && node tests/test-session-naming.mjs && node tests/test-session-connector-naming.mjs && node tests/test-cloudflared-config.mjs && node tests/test-tunnel-diagnostics.mjs && node tests/test-history-index-contract.mjs && node tests/test-session-route-utils.mjs && node tests/test-apps-builtins.mjs && node tests/test-session-app-scope.mjs && node tests/test-session-external-trigger-refresh.mjs && node tests/test-session-follow-up-queue.mjs && node tests/test-session-label-prompt-context.mjs && node tests/test-auto-compaction.mjs && node tests/test-agent-mailbox.mjs && node tests/test-agent-mail-worker.mjs && node tests/test-embedded-mail-worker.mjs && node tests/test-agent-mail-reply.mjs && node tests/test-agent-mail-command.mjs && node tests/test-agent-mail-cloudflare-routing.mjs && node tests/test-mailbox-runtime-registry.mjs && node tests/test-mailbox-ingress-self-check.mjs && node tests/test-connector-feishu.mjs && node tests/test-feishu-connector.mjs && node tests/test-wechat-connector.mjs && node tests/test-whatsapp-business-connector.mjs && node tests/test-connector-surface-public-paths.mjs && node tests/test-connector-surface-health.mjs && node tests/test-guest-wechat-connector-startup.mjs && node tests/test-feishu-ops.mjs && node tests/test-voice-connector.mjs && node tests/test-remote-capability-monitor.mjs && node tests/test-github-ci-auto-repair.mjs && node tests/test-github-ci-auto-repair-runner.mjs && node tests/test-generate-token-preserves-password.mjs && node tests/test-chat-instance-sync.mjs && node tests/test-guest-instance-command.mjs && node tests/test-normalize-owner-instance.mjs && node tests/test-source-runtime-only.mjs && node tests/test-cli-release-removed.mjs && node tests/test-oversized-files-report.mjs && npm run test:merge-safety",
     "test:merge-safety": "node scripts/run-with-clean-instance-env.mjs npm run test:merge-safety:raw",
     "test:merge-safety:raw": "node tests/test-runtime-selection.mjs && node tests/test-chat-tool-default-preference.mjs && node tests/test-tool-visibility.mjs && node tests/test-install-micro-agent.mjs && node tests/test-chat-server-default-micro-agent-bootstrap.mjs && node tests/test-models-micro-agent-per-model-reasoning.mjs && node tests/test-models-codex-fallback.mjs && node tests/test-claude-build-args.mjs && node tests/test-process-runner-runtime-family.mjs && node tests/test-summarizer-managed-codex-home.mjs && node tests/test-detached-assistant-managed-codex-home.mjs && node tests/test-session-pinning.mjs && node tests/test-http-session-patch-runtime-preferences.mjs && node tests/test-http-session-summary-refs.mjs && node tests/test-chat-tooling-layout.mjs && node tests/test-chat-voice-doubao-relay.mjs && node tests/test-chat-upload-button.mjs && node tests/test-chat-share-snapshot-message.mjs && node tests/test-chat-compose-draft.mjs && node tests/test-chat-sidebar-attachments.mjs && node tests/test-chat-frontend-state-boundary.mjs && node tests/test-chat-swipe-gestures.mjs && node tests/test-session-http-sidebar-refresh.mjs && node tests/test-session-http-foreground-refresh.mjs && node tests/test-session-http-ref-hydration.mjs && node tests/test-session-http-share-snapshot-scroll.mjs && node tests/test-chat-ws-reconnect-scroll-suppression.mjs && node tests/test-chat-event-block-append.mjs && node tests/test-chat-event-block-dispatch.mjs && node tests/test-chat-reply-self-check-collapse.mjs && node tests/test-chat-transcript-placeholder-free.mjs && node tests/test-chat-loader-versioned-assets.mjs && node tests/test-chat-build-update-indicator.mjs && node tests/test-chat-split-frontend-smoke.mjs && node tests/test-chat-launch-intent.mjs && node tests/test-source-runtime-only.mjs && node tests/test-source-runtime-prompts.mjs && node tests/test-custom-tool-prompt-shaping.mjs && npm run test:shared-tools && node tests/test-runtime-policy.mjs && node tests/test-system-prompt.mjs && node tests/test-system-prompt-gmail.mjs && node tests/test-chat-nontechnical-copy.mjs && node tests/test-chat-thinking-block-display-preference.mjs && node tests/test-structured-runtime-auth-failure.mjs && node tests/test-user-shell-env.mjs && node tests/test-usage-summary-command.mjs && node tests/test-connector-gmail.mjs",
     "test:shared-tools": "node scripts/run-with-clean-instance-env.mjs npm run test:shared-tools:raw",
