+++ scripts/feishu-connector.mjs
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 
-import { createHash } from 'crypto';
 import { appendFile, mkdir, readFile, rename, rm, writeFile } from 'fs/promises';
 import { homedir } from 'os';
 import { dirname, join, resolve } from 'path';
@@ -15,15 +14,39 @@ import {
 } from '../lib/external-runtime-selection.mjs';
 import { loadMailboxRuntimeRegistry } from '../lib/mailbox-runtime-registry.mjs';
 import {
-  buildAssistantReplyAttachmentFallbackText,
   selectAssistantReplyEvent,
-  stripHiddenBlocks,
 } from '../lib/reply-selection.mjs';
 import { loadUiRuntimeSelection } from '../lib/runtime-selection.mjs';
 import {
   buildConnectorFailureReply,
   decideConnectorUserVisibleReply,
 } from '../lib/connector-user-visible-reply.mjs';
+import {
+  DEFAULT_FEISHU_SESSION_SYSTEM_PROMPT as DEFAULT_SESSION_SYSTEM_PROMPT,
+  FEISHU_CONNECTOR_ID,
+  FEISHU_CONNECTOR_NAME,
+  LARK_CONNECTOR_NAME,
+  LEGACY_DEFAULT_FEISHU_SESSION_SYSTEM_PROMPT as LEGACY_DEFAULT_SESSION_SYSTEM_PROMPT,
+  buildExternalTriggerId,
+  buildFeishuConversationQueueKey,
+  buildFeishuApiUuid,
+  buildFeishuPostContent,
+  buildFeishuTopicId,
+  buildMessageSourceContext,
+  buildRemoteLabMessage,
+  buildRequestId,
+  buildSessionDescription,
+  buildSessionSourceContext,
+  compileFeishuReplyText,
+  getSummaryFeishuImageKeys,
+  isSupportedRemoteLabInboundMessage,
+  normalizeFeishuMode,
+  normalizeReplyText,
+  sanitizeIdPart,
+  summarizeFeishuEvent as summarizeEvent,
+  summarizeFeishuEventForLog as summarizeEventForLog,
+  summarizeFeishuLegacyMessageEvent as summarizeLegacyMessageEvent,
+} from '../connectors/feishu/index.mjs';
 import { ConnectorDriver } from '../lib/connector-driver.mjs';
 import { createFeishuConnectorTransport } from '../lib/connector-driver-transports.mjs';
 import {
@@ -42,31 +65,11 @@ const DEFAULT_ACCESS_STATE_FILENAME = 'access-state.json';
 const DEFAULT_CHAT_BASE_URL = `http://127.0.0.1:${CHAT_PORT}`;
 const DEFAULT_SESSION_TOOL = 'codex';
 const DEFAULT_RUNTIME_SELECTION_MODE = 'ui';
-const LEGACY_DEFAULT_SESSION_SYSTEM_PROMPT = [
-  'You are replying as a Feishu bot powered by RemoteLab on the user\'s own machine.',
-  'For each assistant turn, output exactly the plain-text message to send back to Feishu.',
-  'Keep replies concise, helpful, and natural.',
-  'Match the user\'s language when practical.',
-  'In group chats, prefer silence by default: if the message is mainly human-to-human chatter, laughter, status updates, side remarks, or is not clearly asking for you, output an empty string.',
-  'Reply when you are directly addressed, clearly asked for help or information, asked to take an action, or when a short reply is genuinely useful.',
-  'If the chat asks you to speak less or not reply to every message, treat that as an active local rule until someone clearly changes it.',
-  'If you are unsure whether to reply, choose silence and output an empty string. An empty string means no Feishu message should be sent.',
-  'Do not mention hidden connector, session, or run internals unless the user explicitly asks.',
-].join('\n');
-const DEFAULT_SESSION_SYSTEM_PROMPT = [
-  'You are interacting through a Feishu or Lark bot on the user\'s own machine.',
-  'Keep connector-specific overrides minimal and only describe constraints not already owned by RemoteLab backend prompt logic.',
-].join('\n');
 const RUN_POLL_INTERVAL_MS = 1500;
 const RUN_POLL_TIMEOUT_MS = 0;
-const MAX_FEISHU_TEXT_LENGTH = 5000;
-const MAX_INBOUND_LOG_PREVIEW_LENGTH = 240;
 const MAX_FEISHU_RESOURCE_BYTES = 100 * 1024 * 1024;
 const DEFAULT_PROCESSING_REACTION_EMOJI_TYPE = 'THINKING';
 const CONNECTOR_PID_FILENAME = 'connector.pid';
-const FEISHU_EMOJI_ALIAS_PATTERN = /\[(?:[\u3400-\u9FFF]{1,4})\]/gu;
-const UNICODE_EMOJI_PATTERN = /(?:\p{Regional_Indicator}{2}|[#*0-9]\uFE0F?\u20E3|(?:\p{Extended_Pictographic}|\p{Emoji_Presentation})(?:\uFE0E|\uFE0F)?(?:\u200D(?:\p{Extended_Pictographic}|\p{Emoji_Presentation})(?:\uFE0E|\uFE0F)?)*)/gu;
-const REMOTELAB_SESSION_APP_ID = 'feishu';
 const APPROVE_CURRENT_CHAT_COMMANDS = new Set([
   '授权本群',
   '授权这个群',
@@ -503,343 +506,6 @@ async function releaseConnectorPidLock(lock) {
   await rm(pidPath, { force: true }).catch(() => {});
 }
 
-function parseTextPreview(rawContent) {
-  const content = trimString(rawContent);
-  if (!content) return '';
-  try {
-    const parsed = JSON.parse(content);
-    if (typeof parsed?.text === 'string') {
-      return parsed.text;
-    }
-  } catch {}
-  return '';
-}
-
-function parseMessageContent(rawContent) {
-  const content = trimString(rawContent);
-  if (!content) return null;
-  try {
-    const parsed = JSON.parse(content);
-    if (parsed && typeof parsed === 'object') {
-      return parsed;
-    }
-  } catch {}
-  return null;
-}
-
-function truncateLogPreview(value, maxLength = MAX_INBOUND_LOG_PREVIEW_LENGTH) {
-  const normalized = trimString(value).replace(/\s+/g, ' ');
-  if (!normalized) return '';
-  if (normalized.length <= maxLength) return normalized;
-  return `${normalized.slice(0, maxLength - 1).trimEnd()}…`;
-}
-
-function collectStructuredText(value, fragments, seen, depth = 0) {
-  if (depth > 5 || fragments.length >= 8 || value === null || value === undefined) {
-    return;
-  }
-  if (typeof value === 'string') {
-    const normalized = trimString(value);
-    if (normalized && !seen.has(normalized)) {
-      seen.add(normalized);
-      fragments.push(normalized);
-    }
-    return;
-  }
-  if (Array.isArray(value)) {
-    for (const item of value) {
-      collectStructuredText(item, fragments, seen, depth + 1);
-      if (fragments.length >= 8) return;
-    }
-    return;
-  }
-  if (typeof value !== 'object') return;
-  for (const key of ['title', 'text', 'name', 'label', 'content']) {
-    if (!(key in value)) continue;
-    collectStructuredText(value[key], fragments, seen, depth + 1);
-    if (fragments.length >= 8) return;
-  }
-}
-
-function extractStructuredTextPreview(value) {
-  const fragments = [];
-  collectStructuredText(value, fragments, new Set());
-  return truncateLogPreview(fragments.join(' '));
-}
-
-function selectPostContentVariant(value) {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return null;
-  if (Array.isArray(value.content)) return value;
-  for (const key of ['zh_cn', 'en_us', 'ja_jp']) {
-    const candidate = value[key];
-    if (candidate && typeof candidate === 'object' && Array.isArray(candidate.content)) {
-      return candidate;
-    }
-  }
-  for (const candidate of Object.values(value)) {
-    if (candidate && typeof candidate === 'object' && Array.isArray(candidate.content)) {
-      return candidate;
-    }
-  }
-  return null;
-}
-
-function renderPostElementText(element) {
-  if (typeof element === 'string') return element;
-  if (!element || typeof element !== 'object') return '';
-  const tag = trimString(element.tag).toLowerCase();
-  if (tag === 'at') {
-    const name = trimString(element.user_name || element.name || element.text);
-    const userId = trimString(element.user_id);
-    return name ? `@${name.replace(/^@+/, '')}` : (userId ? `@${userId.replace(/^@+/, '')}` : '');
-  }
-  if (['text', 'md', 'a'].includes(tag)) {
-    return typeof element.text === 'string' ? element.text : '';
-  }
-  if (tag === 'img' || tag === 'image') return '[image]';
-  if (tag === 'media') return '[media]';
-  if (tag === 'file') {
-    const fileName = trimString(element.file_name || element.name);
-    return fileName ? `[file: ${fileName}]` : '[file]';
-  }
-  return typeof element.text === 'string' ? element.text : '';
-}
-
-function extractPostMessageText(parsedContent) {
-  const variant = selectPostContentVariant(parsedContent);
-  if (!variant) return extractStructuredTextPreview(parsedContent);
-
-  const lines = [];
-  const title = trimString(variant.title);
-  if (title) lines.push(title);
-
-  for (const block of variant.content) {
-    const elements = Array.isArray(block) ? block : [block];
-    const line = elements.map(renderPostElementText).join('').trimEnd();
-    if (line.trim()) {
-      lines.push(line);
-    }
-  }
-  return lines.join('\n').trim();
-}
-
-function contentKeyPreview(parsedContent) {
-  if (!parsedContent || Array.isArray(parsedContent) || typeof parsedContent !== 'object') {
-    return [];
-  }
-  return Object.keys(parsedContent).filter(Boolean).slice(0, 6);
-}
-
-function addFeishuImageKey(keys, seen, value) {
-  const key = trimString(value);
-  if (!key || seen.has(key)) return;
-  seen.add(key);
-  keys.push(key);
-}
-
-function collectFeishuImageKeys(value, keys, seen, depth = 0) {
-  if (depth > 8 || value === null || value === undefined) {
-    return;
-  }
-  if (Array.isArray(value)) {
-    for (const item of value) {
-      collectFeishuImageKeys(item, keys, seen, depth + 1);
-    }
-    return;
-  }
-  if (typeof value !== 'object') {
-    return;
-  }
-
-  addFeishuImageKey(keys, seen, value.image_key);
-  addFeishuImageKey(keys, seen, value.imageKey);
-  const tag = trimString(value.tag).toLowerCase();
-  if (tag === 'img' || tag === 'image') {
-    addFeishuImageKey(keys, seen, value.file_key);
-    addFeishuImageKey(keys, seen, value.fileKey);
-  }
-  for (const child of Object.values(value)) {
-    collectFeishuImageKeys(child, keys, seen, depth + 1);
-  }
-}
-
-function extractFeishuImageKeysFromContent(parsedContent) {
-  const keys = [];
-  collectFeishuImageKeys(parsedContent, keys, new Set());
-  return keys;
-}
-
-function summarizeMessageContent(messageType, rawContent) {
-  const normalizedType = trimString(messageType).toLowerCase();
-  const parsedContent = parseMessageContent(rawContent);
-  const imageKeys = extractFeishuImageKeysFromContent(parsedContent);
-  let messageText = '';
-  let textPreview = '';
-
-  if (normalizedType === 'text') {
-    messageText = parseTextPreview(rawContent) || trimString(rawContent);
-    textPreview = messageText;
-  } else if (normalizedType === 'post') {
-    messageText = extractPostMessageText(parsedContent);
-    textPreview = truncateLogPreview(messageText || extractStructuredTextPreview(parsedContent));
-  } else if (normalizedType === 'file') {
-    textPreview = trimString(parsedContent?.file_name || parsedContent?.name);
-  } else if (normalizedType === 'share_chat') {
-    textPreview = trimString(parsedContent?.chat_name || parsedContent?.name || parsedContent?.chat_id);
-  } else if (normalizedType === 'share_user') {
-    textPreview = trimString(parsedContent?.user_name || parsedContent?.name || parsedContent?.user_id);
-  } else if (normalizedType === 'location') {
-    textPreview = trimString(parsedContent?.name || parsedContent?.title || parsedContent?.address);
-  } else if (normalizedType === 'interactive') {
-    textPreview = extractStructuredTextPreview(parsedContent);
-  }
-
-  const contentSummary = (() => {
-    switch (normalizedType) {
-      case 'text':
-        return textPreview ? `Text message: ${truncateLogPreview(textPreview)}` : 'Text message';
-      case 'image':
-        return 'Image attachment';
-      case 'file':
-        return textPreview ? `File attachment: ${truncateLogPreview(textPreview)}` : 'File attachment';
-      case 'audio':
-        return 'Audio attachment';
-      case 'media':
-        return 'Media attachment';
-      case 'sticker':
-        return 'Sticker message';
-      case 'post':
-        return textPreview ? `Rich text post: ${truncateLogPreview(textPreview)}` : 'Rich text post';
-      case 'share_chat':
-        return textPreview ? `Shared chat: ${truncateLogPreview(textPreview)}` : 'Shared chat';
-      case 'share_user':
-        return textPreview ? `Shared contact: ${truncateLogPreview(textPreview)}` : 'Shared contact';
-      case 'location':
-        return textPreview ? `Location message: ${truncateLogPreview(textPreview)}` : 'Location message';
-      case 'interactive':
-        return textPreview ? `Interactive card: ${truncateLogPreview(textPreview)}` : 'Interactive card';
-      default: {
-        const typeLabel = normalizedType || 'unknown';
-        const keys = contentKeyPreview(parsedContent);
-        return keys.length
-          ? `Unsupported message (${typeLabel}; keys=${keys.join(',')})`
-          : `Unsupported message (${typeLabel})`;
-      }
-    }
-  })();
-
-  return {
-    messageText,
-    textPreview,
-    contentSummary,
-    contentKeys: contentKeyPreview(parsedContent),
-    imageKeys,
-  };
-}
-
-function summarizeEventForLog(summary) {
-  return {
-    eventId: summary?.eventId || '',
-    eventType: summary?.eventType || '',
-    chatId: summary?.chatId || '',
-    chatType: summary?.chatType || '',
-    groupMessageType: summary?.groupMessageType || '',
-    chatMode: summary?.chatMode || '',
-    messageId: summary?.messageId || '',
-    messageType: summary?.messageType || '',
-    rootId: summary?.rootId || '',
-    threadId: summary?.threadId || '',
-    senderOpenId: summary?.sender?.openId || '',
-    mentionCount: Array.isArray(summary?.mentions) ? summary.mentions.length : 0,
-    textPreview: truncateLogPreview(summary?.textPreview),
-    contentSummary: truncateLogPreview(summary?.contentSummary),
-  };
-}
-
-function summarizeEvent(data) {
-  const sender = data?.sender || {};
-  const senderId = sender?.sender_id || {};
-  const message = data?.message || {};
-  const mentions = Array.isArray(message.mentions) ? message.mentions : [];
-  const rawContent = typeof message.content === 'string' ? message.content : '';
-  const normalizedContent = summarizeMessageContent(message.message_type || '', rawContent);
-  return {
-    eventId: data?.event_id || '',
-    eventType: data?.event_type || '',
-    tenantKey: data?.tenant_key || '',
-    appId: data?.app_id || '',
-    createTime: data?.create_time || '',
-    sender: {
-      openId: senderId?.open_id || '',
-      userId: senderId?.user_id || '',
-      unionId: senderId?.union_id || '',
-      senderType: sender?.sender_type || '',
-      tenantKey: sender?.tenant_key || '',
-    },
-    chatId: message.chat_id || '',
-    chatType: message.chat_type || '',
-    groupMessageType: message.group_message_type || data?.group_message_type || '',
-    chatMode: message.chat_mode || data?.chat_mode || '',
-    messageId: message.message_id || '',
-    rootId: message.root_id || '',
-    parentId: message.parent_id || '',
-    threadId: message.thread_id || '',
-    messageType: message.message_type || '',
-    mentions: mentions.map((mention) => ({
-      key: mention?.key || '',
-      name: mention?.name || '',
-      openId: mention?.id?.open_id || '',
-      userId: mention?.id?.user_id || '',
-      unionId: mention?.id?.union_id || '',
-      tenantKey: mention?.tenant_key || '',
-    })),
-    messageText: normalizedContent.messageText,
-    textPreview: normalizedContent.textPreview,
-    contentSummary: normalizedContent.contentSummary,
-    contentKeys: normalizedContent.contentKeys,
-    imageKeys: normalizedContent.imageKeys,
-    rawContent,
-  };
-}
-
-function summarizeLegacyMessageEvent(data) {
-  const messageType = data?.msg_type || data?.message_type || '';
-  const rawContent = typeof data?.text === 'string' ? data.text : '';
-  const normalizedContent = summarizeMessageContent(messageType, rawContent);
-  return {
-    eventId: data?.uuid || data?.event_id || '',
-    eventType: 'message',
-    tenantKey: data?.tenant_key || '',
-    appId: data?.app_id || '',
-    createTime: data?.ts || '',
-    sender: {
-      openId: data?.open_id || data?.sender?.open_id || '',
-      userId: data?.employee_id || data?.sender?.employee_id || '',
-      unionId: '',
-      senderType: 'user',
-      tenantKey: data?.tenant_key || '',
-    },
-    chatId: data?.open_chat_id || data?.chat_id || '',
-    chatType: data?.chat_type || '',
-    groupMessageType: data?.group_message_type || '',
-    chatMode: data?.chat_mode || '',
-    messageId: data?.open_message_id || data?.message_id || '',
-    rootId: '',
-    parentId: '',
-    threadId: '',
-    messageType,
-    mentions: [],
-    messageText: typeof data?.text_without_at_bot === 'string'
-      ? data.text_without_at_bot
-      : normalizedContent.messageText,
-    textPreview: typeof data?.text_without_at_bot === 'string' ? data.text_without_at_bot : normalizedContent.textPreview,
-    contentSummary: normalizedContent.contentSummary,
-    contentKeys: normalizedContent.contentKeys,
-    imageKeys: normalizedContent.imageKeys,
-    rawContent,
-  };
-}
-
 async function ensureDir(pathname) {
   await mkdir(pathname, { recursive: true });
 }
@@ -1141,217 +807,8 @@ async function recordInboundEvent(runtime, summary, raw, sourceLabel) {
   return allowed;
 }
 
-function containsCjk(text) {
-  return /[\u3400-\u9fff\uf900-\ufaff]/u.test(text || '');
-}
-
-function sanitizeIdPart(value) {
-  return String(value || '')
-    .trim()
-    .replace(/[^a-zA-Z0-9._:-]+/g, '_')
-    .replace(/_+/g, '_')
-    .replace(/^_+|_+$/g, '') || 'unknown';
-}
-
-function normalizeFeishuMode(value) {
-  return trimString(value).toLowerCase().replace(/[_\s-]+/g, '_');
-}
-
-function isFeishuTopicChat(summary) {
-  const chatType = normalizeFeishuMode(summary?.chatType);
-  const groupMessageType = normalizeFeishuMode(summary?.groupMessageType);
-  const chatMode = normalizeFeishuMode(summary?.chatMode);
-  return chatType === 'topic'
-    || groupMessageType === 'thread'
-    || groupMessageType === 'topic'
-    || chatMode === 'thread'
-    || chatMode === 'topic';
-}
-
-function buildFeishuTopicId(summary) {
-  const threadId = trimString(summary?.threadId);
-  if (threadId) return threadId;
-
-  const rootId = trimString(summary?.rootId);
-  if (rootId) return rootId;
-
-  if (isFeishuTopicChat(summary)) {
-    return trimString(summary?.parentId) || trimString(summary?.messageId);
-  }
-
-  return '';
-}
-
-function buildFeishuConversationKind(summary) {
-  return buildFeishuTopicId(summary) ? 'topic' : sanitizeIdPart(summary?.chatType || 'chat');
-}
-
-function buildExternalTriggerId(summary) {
-  const chatId = sanitizeIdPart(summary?.chatId || 'unknown_chat');
-  const topicId = buildFeishuTopicId(summary);
-  if (topicId) {
-    return `feishu:topic:${chatId}:${sanitizeIdPart(topicId)}`;
-  }
-  return `feishu:${sanitizeIdPart(summary?.chatType || 'chat')}:${chatId}`;
-}
-
-function buildRequestId(summary) {
-  return `feishu:${sanitizeIdPart(summary.messageId || `${Date.now()}`)}`;
-}
-
-function buildReplyUuid(summary) {
-  return `reply:${sanitizeIdPart(summary.messageId || `${Date.now()}`).slice(0, 60)}`;
-}
-
-function buildFeishuApiUuid(value, summary) {
-  const seed = trimString(value) || buildReplyUuid(summary);
-  const normalized = seed
-    .replace(/[^a-zA-Z0-9_-]+/g, '_')
-    .replace(/_+/g, '_')
-    .replace(/^_+|_+$/g, '');
-  if (normalized && normalized.length <= 64) {
-    return normalized;
-  }
-  const digest = createHash('sha1').update(seed || `${Date.now()}`).digest('hex').slice(0, 32);
-  return `rl_${digest}`;
-}
-
 function buildSessionName() {
   return '';
-}
-
-function buildSessionDescription(summary) {
-  if (buildFeishuTopicId(summary)) {
-    return 'Inbound Feishu topic thread';
-  }
-  const chatType = trimString(summary?.chatType);
-  return chatType ? `Inbound Feishu ${chatType} chat` : 'Inbound Feishu chat';
-}
-
-function mentionDisplayName(mention) {
-  const name = trimString(mention?.name);
-  if (name) return name;
-  const token = trimString(mention?.key).replace(/^@+/, '');
-  return token || 'user';
-}
-
-function renderMentionPreview(text, mentions) {
-  let rendered = trimString(text);
-  if (!rendered) return '';
-  for (const mention of Array.isArray(mentions) ? mentions : []) {
-    const token = trimString(mention?.key);
-    if (!token) continue;
-    rendered = rendered.split(token).join(`@${mentionDisplayName(mention)}`);
-  }
-  return rendered;
-}
-
-function getSummaryFeishuImageKeys(summary) {
-  const explicitKeys = Array.isArray(summary?.imageKeys)
-    ? summary.imageKeys.map((key) => trimString(key)).filter(Boolean)
-    : [];
-  if (explicitKeys.length > 0) {
-    return Array.from(new Set(explicitKeys));
-  }
-  return extractFeishuImageKeysFromContent(parseMessageContent(summary?.rawContent));
-}
-
-function isSupportedRemoteLabInboundMessage(summary) {
-  const messageType = trimString(summary?.messageType).toLowerCase();
-  if (!messageType || messageType === 'text') return true;
-  if (messageType === 'image') return getSummaryFeishuImageKeys(summary).length > 0;
-  if (messageType === 'post') {
-    return Boolean(trimString(summary?.messageText || summary?.textPreview))
-      || getSummaryFeishuImageKeys(summary).length > 0;
-  }
-  return false;
-}
-
-function buildRemoteLabMessage(summary) {
-  const rawMessage = trimString(summary.messageText) || trimString(summary.textPreview);
-  const renderedMessage = renderMentionPreview(rawMessage, summary.mentions);
-  const displayMessage = renderedMessage || rawMessage || trimString(summary.contentSummary) || '[non-text or empty message]';
-  const senderName = trimString(summary?.sender?.name || summary?.sender?.displayName);
-  const senderPrefix = summary?.chatType === 'group' && senderName ? `${senderName}: ` : '';
-  const downloadFailures = Array.isArray(summary?.attachmentDownloadFailures)
-    ? summary.attachmentDownloadFailures
-    : [];
-  const failureText = downloadFailures.length > 0
-    ? `\n\n[Feishu attachment download failed for ${downloadFailures.length} image(s).]`
-    : '';
-  return `${senderPrefix}${displayMessage}${failureText}`;
-}
-
-function buildSessionSourceContext(summary) {
-  const topicId = buildFeishuTopicId(summary);
-  const context = {
-    connector: 'feishu',
-    conversationKind: buildFeishuConversationKind(summary),
-    chatType: trimString(summary?.chatType),
-    chatId: trimString(summary?.chatId),
-  };
-  const chatName = trimString(summary?.chatName);
-  if (chatName) context.chatName = chatName;
-  const groupMessageType = trimString(summary?.groupMessageType);
-  if (groupMessageType) context.groupMessageType = groupMessageType;
-  const chatMode = trimString(summary?.chatMode);
-  if (chatMode) context.chatMode = chatMode;
-  if (topicId) context.topicId = topicId;
-  const threadId = trimString(summary?.threadId);
-  if (threadId) context.threadId = threadId;
-  const rootId = trimString(summary?.rootId);
-  if (rootId) context.rootId = rootId;
-  return context;
-}
-
-function buildMessageSourceContext(summary) {
-  const topicId = buildFeishuTopicId(summary);
-  const imageKeys = getSummaryFeishuImageKeys(summary);
-  const context = {
-    connector: 'feishu',
-    messageId: trimString(summary?.messageId),
-    chatType: trimString(summary?.chatType),
-    conversationKind: buildFeishuConversationKind(summary),
-  };
-  if (topicId) context.topicId = topicId;
-  const threadId = trimString(summary?.threadId);
-  if (threadId) context.threadId = threadId;
-  const rootId = trimString(summary?.rootId);
-  if (rootId) context.rootId = rootId;
-  const parentId = trimString(summary?.parentId);
-  if (parentId) context.parentId = parentId;
-  const groupMessageType = trimString(summary?.groupMessageType);
-  if (groupMessageType) context.groupMessageType = groupMessageType;
-  const chatMode = trimString(summary?.chatMode);
-  if (chatMode) context.chatMode = chatMode;
-  const senderName = trimString(summary?.sender?.name || summary?.sender?.displayName);
-  if (senderName) {
-    context.sender = { name: senderName };
-  }
-  const mentions = (Array.isArray(summary?.mentions) ? summary.mentions : [])
-    .map((mention) => {
-      const name = mentionDisplayName(mention);
-      const token = trimString(mention?.key);
-      if (!name && !token) return null;
-      return {
-        ...(name ? { name } : {}),
-        ...(token ? { token } : {}),
-      };
-    })
-    .filter(Boolean);
-  if (mentions.length > 0) {
-    context.mentions = mentions;
-  }
-  const contentSummary = trimString(summary?.contentSummary);
-  if (contentSummary) {
-    context.contentSummary = contentSummary;
-  }
-  if (imageKeys.length > 0) {
-    context.attachments = {
-      imageCount: imageKeys.length,
-    };
-  }
-  return context;
 }
 
 function normalizeHeaderValue(headers, name) {
@@ -1590,27 +1047,6 @@ async function loadAssistantReply(requester, sessionId, runId, requestId) {
   return candidate;
 }
 
-function normalizeReplyText(text) {
-  const normalized = stripOutboundEmojiArtifacts(stripHiddenBlocks(String(text || '').replace(/\r\n/g, '\n'))).trim();
-  if (!normalized) return '';
-  if (normalized.length <= MAX_FEISHU_TEXT_LENGTH) return normalized;
-  return `${normalized.slice(0, MAX_FEISHU_TEXT_LENGTH - 16).trimEnd()}\n\n[truncated]`;
-}
-
-function stripOutboundEmojiArtifacts(text) {
-  return String(text || '')
-    .replace(UNICODE_EMOJI_PATTERN, ' ')
-    .replace(FEISHU_EMOJI_ALIAS_PATTERN, ' ')
-    .replace(/[\u200D\uFE0E\uFE0F]/g, '')
-    .replace(/\s+([,.;:!?，。！？；：、])/g, '$1')
-    .replace(/([([{（【])\s+/g, '$1')
-    .replace(/\s+([)\]}）】])/g, '$1')
-    .replace(/[ \t]+\n/g, '\n')
-    .replace(/\n[ \t]+/g, '\n')
-    .replace(/[ \t]{2,}/g, ' ')
-    .replace(/\n{3,}/g, '\n\n');
-}
-
 function isMainModule() {
   if (!process.argv[1]) return false;
   return import.meta.url === pathToFileURL(resolve(process.argv[1])).href;
@@ -1682,8 +1118,8 @@ function createRuntimeContext(config, storagePaths, accessState) {
   };
 }
 
-function enqueueByChat(runtime, summary, worker) {
-  const key = summary.chatId || summary.messageId || 'unknown_chat';
+function enqueueByConversation(runtime, summary, worker) {
+  const key = buildFeishuConversationQueueKey(summary);
   const previous = runtime.chatQueues.get(key) || Promise.resolve();
   const next = previous
     .catch(() => {})
@@ -1913,9 +1349,9 @@ async function createOrReuseSession(runtime, summary, runtimeSelection) {
     folder: runtime.config.sessionFolder,
     tool: runtimeSelection.tool,
     name: buildSessionName(summary),
-    sourceId: 'feishu',
+    sourceId: FEISHU_CONNECTOR_ID,
     sourceName,
-    group: 'Feishu',
+    group: FEISHU_CONNECTOR_NAME,
     description: buildSessionDescription(summary),
     systemPrompt: runtime.config.systemPrompt,
     externalTriggerId: buildExternalTriggerId(summary),
@@ -2034,9 +1470,9 @@ async function generateRemoteLabReply(runtime, summary) {
     folder: runtime.config.sessionFolder,
     tool: runtimeSelection.tool,
     name: buildSessionName(effectiveSummary),
-    sourceId: 'feishu',
-    sourceName: runtime.config.region === 'lark-global' ? 'Lark' : 'Feishu',
-    group: 'Feishu',
+    sourceId: FEISHU_CONNECTOR_ID,
+    sourceName: runtime.config.region === 'lark-global' ? LARK_CONNECTOR_NAME : FEISHU_CONNECTOR_NAME,
+    group: FEISHU_CONNECTOR_NAME,
     description: buildSessionDescription(effectiveSummary),
     systemPrompt: runtime.config.systemPrompt,
     externalTriggerId: buildExternalTriggerId(effectiveSummary),
@@ -2086,107 +1522,6 @@ async function generateRemoteLabReply(runtime, summary) {
     replyText,
     silent: !replyText,
   };
-}
-
-function resolveMentionTargetId(mention) {
-  return trimString(mention?.openId) || trimString(mention?.userId) || trimString(mention?.unionId);
-}
-
-function escapeFeishuMentionValue(value) {
-  return String(value || '')
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;');
-}
-
-function compileFeishuReplyText(text, mentions) {
-  let compiled = normalizeReplyText(text);
-  const normalizedMentions = (Array.isArray(mentions) ? mentions : [])
-    .map((mention) => ({
-      mention,
-      token: trimString(mention?.key),
-      targetId: resolveMentionTargetId(mention),
-      displayName: mentionDisplayName(mention),
-    }))
-    .filter((entry) => entry.targetId && (entry.token || entry.displayName));
-  for (const { mention, token, targetId } of normalizedMentions.sort((left, right) => right.token.length - left.token.length)) {
-    if (!token) continue;
-    const tag = `<at user_id="${escapeFeishuMentionValue(targetId)}">${escapeFeishuMentionValue(mentionDisplayName(mention))}</at>`;
-    compiled = compiled.split(token).join(tag);
-  }
-  for (const { mention, displayName, targetId } of normalizedMentions.sort((left, right) => right.displayName.length - left.displayName.length)) {
-    if (!displayName) continue;
-    const alias = `@${displayName}`;
-    const tag = `<at user_id="${escapeFeishuMentionValue(targetId)}">${escapeFeishuMentionValue(mentionDisplayName(mention))}</at>`;
-    compiled = compiled.split(alias).join(tag);
-  }
-  return compiled;
-}
-
-function normalizeMentionEntries(mentions) {
-  return (Array.isArray(mentions) ? mentions : [])
-    .map((mention) => ({
-      mention,
-      token: trimString(mention?.key),
-      targetId: resolveMentionTargetId(mention),
-      displayName: mentionDisplayName(mention),
-    }))
-    .filter((entry) => entry.targetId && (entry.token || entry.displayName));
-}
-
-function findNextMentionMatch(line, mentionEntries, startIndex) {
-  let best = null;
-  for (const entry of mentionEntries) {
-    const markers = [
-      entry.token,
-      entry.displayName ? `@${entry.displayName}` : '',
-    ].filter(Boolean);
-    for (const marker of markers) {
-      const index = line.indexOf(marker, startIndex);
-      if (index < 0) continue;
-      if (!best || index < best.index || (index === best.index && marker.length > best.marker.length)) {
-        best = { entry, marker, index };
-      }
-    }
-  }
-  return best;
-}
-
-function renderFeishuMarkdownLineElements(line, mentionEntries) {
-  const elements = [];
-  let cursor = 0;
-  while (cursor < line.length) {
-    const match = findNextMentionMatch(line, mentionEntries, cursor);
-    if (!match) {
-      elements.push({ tag: 'md', text: line.slice(cursor) });
-      break;
-    }
-    if (match.index > cursor) {
-      elements.push({ tag: 'md', text: line.slice(cursor, match.index) });
-    }
-    elements.push({
-      tag: 'at',
-      user_id: match.entry.targetId,
-      user_name: match.entry.displayName || match.entry.targetId,
-    });
-    cursor = match.index + match.marker.length;
-  }
-  return elements.filter((element) => element.text !== '');
-}
-
-function buildFeishuPostContent(text, mentions) {
-  const normalized = normalizeReplyText(text);
-  const mentionEntries = normalizeMentionEntries(mentions)
-    .sort((left, right) => Math.max(right.token.length, right.displayName.length) - Math.max(left.token.length, left.displayName.length));
-  const content = normalized.split('\n').map((line) => {
-    if (!line.trim()) {
-      return [{ tag: 'text', text: '\u200B' }];
-    }
-    const elements = renderFeishuMarkdownLineElements(line, mentionEntries);
-    return elements.length > 0 ? elements : [{ tag: 'text', text: '\u200B' }];
-  });
-  return JSON.stringify({ zh_cn: { content } });
 }
 
 function isProcessingReactionEnabled(runtime) {
@@ -2711,20 +2046,20 @@ async function main() {
       const summary = summarizeEvent(data);
       const allowed = await recordInboundEvent(runtime, summary, data, 'im.message.receive_v1');
       if (allowed) {
-        enqueueByChat(runtime, summary, () => handleMessage(runtime, summary, 'im.message.receive_v1'));
+        enqueueByConversation(runtime, summary, () => handleMessage(runtime, summary, 'im.message.receive_v1'));
       }
       return {};
     },
     'im.chat.member.user.added_v1': async (data) => {
       const summary = summarizeChatMemberUserAddedEvent(data);
-      enqueueByChat(runtime, { chatId: summary.chatId, messageId: summary.eventId }, () => handleChatMemberUserAdded(runtime, summary, data, 'im.chat.member.user.added_v1'));
+      enqueueByConversation(runtime, { chatId: summary.chatId, messageId: summary.eventId }, () => handleChatMemberUserAdded(runtime, summary, data, 'im.chat.member.user.added_v1'));
       return {};
     },
     message: async (data) => {
       const summary = summarizeLegacyMessageEvent(data);
       const allowed = await recordInboundEvent(runtime, summary, data, 'message');
       if (allowed) {
-        enqueueByChat(runtime, summary, () => handleMessage(runtime, summary, 'message'));
+        enqueueByConversation(runtime, summary, () => handleMessage(runtime, summary, 'message'));
       }
       return {};
     },
