+++ static/chat/i18n.js
@@ -76,6 +76,8 @@
       "sidebar.filter.admin": "Admin ({count})",
       "sidebar.filter.userCount": "{name} ({count})",
       "sidebar.filter.source.chat": "Chat UI",
+      "sidebar.filter.source.feishu": "Feishu",
+      "sidebar.filter.source.email": "Email",
       "sidebar.filter.source.bots": "Bots",
       "sidebar.filter.source.automation": "Automation",
       "sidebar.pinned": "Pinned",
@@ -483,6 +485,8 @@
       "sidebar.filter.admin": "管理员 ({count})",
       "sidebar.filter.userCount": "{name} ({count})",
       "sidebar.filter.source.chat": "聊天界面",
+      "sidebar.filter.source.feishu": "飞书",
+      "sidebar.filter.source.email": "邮件",
       "sidebar.filter.source.bots": "机器人",
       "sidebar.filter.source.automation": "自动化",
       "sidebar.pinned": "置顶",
