+++ static/chat/session-http.js
@@ -237,11 +237,15 @@ const SESSION_LIST_ORGANIZER_POLL_TIMEOUT_MS = 90 * 1000;
 const SESSION_LIST_ORGANIZER_INTERNAL_ROLE = "session_list_organizer";
 const DEFAULT_SORT_SESSION_LIST_BUTTON_LABEL = "Sort List";
 const SESSION_HTTP_SOURCE_FILTER_CHAT_VALUE = typeof SOURCE_FILTER_CHAT_VALUE !== "undefined" ? SOURCE_FILTER_CHAT_VALUE : "chat_ui";
+const SESSION_HTTP_SOURCE_FILTER_FEISHU_VALUE = typeof SOURCE_FILTER_FEISHU_VALUE !== "undefined" ? SOURCE_FILTER_FEISHU_VALUE : "feishu";
+const SESSION_HTTP_SOURCE_FILTER_EMAIL_VALUE = typeof SOURCE_FILTER_EMAIL_VALUE !== "undefined" ? SOURCE_FILTER_EMAIL_VALUE : "email";
 const SESSION_HTTP_SOURCE_FILTER_BOT_VALUE = typeof SOURCE_FILTER_BOT_VALUE !== "undefined" ? SOURCE_FILTER_BOT_VALUE : "bot";
 const SESSION_HTTP_SOURCE_FILTER_AUTOMATION_VALUE = typeof SOURCE_FILTER_AUTOMATION_VALUE !== "undefined" ? SOURCE_FILTER_AUTOMATION_VALUE : "automation";
 const SESSION_HTTP_FILTER_ALL_VALUE = typeof FILTER_ALL_VALUE !== "undefined" ? FILTER_ALL_VALUE : "all";
 const SESSION_LIST_ORGANIZER_SOURCE_LABELS = {
   [SESSION_HTTP_SOURCE_FILTER_CHAT_VALUE]: "Chat UI",
+  [SESSION_HTTP_SOURCE_FILTER_FEISHU_VALUE]: "Feishu",
+  [SESSION_HTTP_SOURCE_FILTER_EMAIL_VALUE]: "Email",
   [SESSION_HTTP_SOURCE_FILTER_BOT_VALUE]: "Bot",
   [SESSION_HTTP_SOURCE_FILTER_AUTOMATION_VALUE]: "Automation",
   [SESSION_HTTP_FILTER_ALL_VALUE]: "All origins",
